+++ concordium-std/src/lib.rs
@@ -235,7 +235,7 @@ fn abort_panic(_info: &core::panic::PanicInfo) -> ! {
 /// Re-export.
 #[cfg(not(feature = "std"))]
 pub use alloc::{
-    borrow::ToOwned, boxed, boxed::Box, rc, string, string::String, string::ToString, vec, vec::Vec,
+    format, borrow::ToOwned, boxed, boxed::Box, rc, string, string::String, string::ToString, vec, vec::Vec,
 };
 /// Re-export.
 #[cfg(not(feature = "std"))]