+++ concordium-std/CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Unreleased changes
 
+## concordium-std 5.1.0 (2022-12-14)
+
 - Add a new primitive `get_random` for generating random numbers in Wasm code testing; `get_random` can be used in tests only, not available for smart contracts on the chain.
 - Fix a linking issue when compiling contracts to native code on Windows and OSX.